Add initial implementation (#1)
This commit builds up the majority of the action: when a given command
is run, the action will check if there's any diff output in the cloned
repo. If there is, it will report a job summary of what changed.

An older implementation of this used annotations, rather than job
summaries, but these don't really work that well, as we can't comment on
files which have changed outside of those referenced in the diff, so the
output is mostly shown in the log output or the autogenerated job
summary. Instead, we assemble a custom output, which is much nicer.

For testing, we use `jet_black`, which allows us to run "black box"
tests in a clean environment using RSpec which is quite pleasant to work
with. To be able to do this, `diff-check` is a script that's called from
the `action.yml` and the expectation is that all the logic should be in
there.

https://github.blog/2022-05-09-supercharging-github-actions-with-job-summaries/
https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#adding-a-job-summary
https://github.com/orgs/community/discussions/9099
https://github.com/odlp/jet_black