Fix path to diff-check in action.yml
It was previously assumed that `$GITHUB_PATH` affected the `$PATH`
lookup but this wasn't actually the case and so the command could never
be run.

Instead, this uses `$GITHUB_ACTION_PATH` (which is the path to the
action cached on the runner) to built out a full path for `diff-check`
instead, which is also simpler.

Fixes #4.