Switch from diff-index to diff
`diff-index`:

    Compare the content and mode of the blobs found in a tree object
    with the corresponding tracked files in the working tree, or with
    the corresponding paths in the index.

Unlike `diff`, which only compares the content. This has the implication
(only on Linux) that if you were to recreate a file (or do something
like change a file's timestamp with `touch`, as in the test) then a
check with `diff-index` would drop us into that code path for there to
subsequently be no known files which had changed.

This behaviour came up when testing with Appraisal, which re-creates
lockfiles when run, but the content is the same.

https://stackoverflow.com/questions/24197606/whats-the-difference-between-git-diff-and-gif-diff-index
https://github.com/thoughtbot/appraisal